etl/etl-requirements.txt
psycopg2-binary
numpy
scikit-learn
pandas
pytest